+++ src/createStoreRegistry.ts
@@ -7,7 +7,7 @@ import type { LoadingStateValue } from "./loadingState";
 import type { Config, ValidatedConfig } from "./config";
 
 export type CrudState<T, S> = {
-  list: { [key: string]: T } | null;
+  record: { [key: string]: T } | null;
   count: number;
   setList: (data: T[]) => void;
   setCount: (count: number) => void;
@@ -30,12 +30,12 @@ export type CrudStore<
   config: V;
 };
 
-export function createCrudStoreRegistry<Models extends Record<string, any>>() {
+export function createStoreRegistry<Models extends Record<string, any>>() {
   const storeRegistry: {
     [K in keyof Models]?: Record<string, any> ;
   } = {};
 
-  function getOrCreateCrudStore<
+  function getOrCreateStore<
     K extends Extract<keyof Models, string>,
     C extends Config<K, Models[K]>,
     V extends ValidatedConfig<K, Models[K], C>
@@ -49,16 +49,16 @@ export function createCrudStoreRegistry<Models extends Record<string, any>>() {
 
       const store: CrudStore<Models[K], K, C, typeof validated> = Object.assign(
         create<CrudState<Models[K], C['state']>>((set) => ({
-          list: null,
+          record: null,
           count: 0,
-          setList: (data) => set({ list: Object.fromEntries(data.map((obj) => [obj[byKey], obj]))}),
+          setList: (data) => set({ record: Object.fromEntries(data.map((obj) => [obj[byKey], obj]))}),
           setInstance: (instance: Models[K], incrementCount?: boolean) =>
             set((state) => {
-              if (!state.list) return {};
+              if (!state.record) return {};
 
               return {
-                list: {
-                  ...state.list,
+                record: {
+                  ...state.record,
                   [instance[byKey]]: instance,
                 },
                 ...incrementCount ? { count: state.count + 1 } : {}
@@ -66,13 +66,13 @@ export function createCrudStoreRegistry<Models extends Record<string, any>>() {
             }),
           updateInstance: (instance: Models[K]) =>
             set((state) => {
-              if (!state.list) return {};
+              if (!state.record) return {};
 
               return {
-                list: {
-                  ...state.list,
+                record: {
+                  ...state.record,
                   [instance[byKey]]: {
-                    ...state.list[byKey] || {},
+                    ...state.record[byKey] || {},
                     ...instance,
                   },
                 },
@@ -80,15 +80,15 @@ export function createCrudStoreRegistry<Models extends Record<string, any>>() {
             }),
           deleteInstance: (instance: Models[K]) =>
             set((state) => {
-              if (!state.list) return {};
+              if (!state.record) return {};
 
-              const newList = { ...state.list };
+              const newList = { ...state.record };
               if (newList[instance[byKey as string]]) {
                 delete newList[instance[byKey as string]];
               }
 
               return {
-                list: newList,
+                record: newList,
                 count: state.count > 0 ? state.count - 1 : 0,
               };
             }),
@@ -123,15 +123,13 @@ export function createCrudStoreRegistry<Models extends Record<string, any>>() {
 
     return storeRegistry[key] as CrudStore<Models[K], K, C, V>;
   }
-  return {
-    getOrCreateCrudStore,
-  } as {
-    getOrCreateCrudStore: <
+  return getOrCreateStore as (
+    <
       K extends Extract<keyof Models, string>,
       C extends Config<K, Models[K]>
     >(
       key: K,
       config: C
-    ) => CrudStore<Models[K], K, C, ValidatedConfig<K, Models[K], C>>;
-  };
+    ) => CrudStore<Models[K], K, C, ValidatedConfig<K, Models[K], C>>
+  );
 };