+++ change-log.md
@@ -4,3 +4,7 @@
 
 ##### Version 0.0.2
 - Include key in `Config` type
+
+##### Version 0.0.3
+- Refactor imports names
+- New `includeRecord` in config to output full `record` in hook